Improved factoring around integrati0on framework boundaries and sync.
