+++ src/hi/integrations/templates/integrations/panes/sync_actions.html
@@ -1,0 +1,26 @@
+{% load icons %}
+{% comment %}
+  Shared IMPORT / REFRESH framework action.
+
+  Sync is a framework concern; integrations participate by providing an
+  IntegrationSynchronizer from their gateway's get_synchronizer().
+  Rendering is conditional on (a) the integration being enabled and
+  (b) the gateway returning a synchronizer. Integrations whose gateway
+  returns None (sync is opt-in) get no button.
+
+  Label flips on `core.has_entities`: IMPORT for the first-time path,
+  REFRESH for subsequent re-syncs. The framework manage view computes
+  has_entities into the `core` context.
+{% endcomment %}
+{% if core.integration_data.is_enabled and core.integration_data.integration_gateway.get_synchronizer %}
+<a href="{% url 'integrations_pre_sync' integration_id=core.integration_data.integration_id %}"
+   class="btn {% if core.has_entities %}btn-outline-secondary{% else %}btn-outline-primary{% endif %}"
+   role="button"
+   data-async="modal">
+  {% if core.has_entities %}
+  {% icon "sync" size="sm" css_class="hi-icon-left" %}REFRESH
+  {% else %}
+  {% icon "upload" size="sm" css_class="hi-icon-left" %}IMPORT
+  {% endif %}
+</a>
+{% endif %}
