+++ src/hi/integrations/tests/test_entity_operations.py
@@ -3,7 +3,8 @@ Unit tests for EntityIntegrationOperations.
 
 Only covers behavior that encodes real classification / transformation /
 graph-traversal logic. The preserve_with_user_data path is already tested
-indirectly via test_sync_mixins (which exercises _remove_entity_intelligently).
+indirectly via test_integration_synchronizer (which exercises
+_remove_entity_intelligently).
 """
 
 import logging
