+++ src/hi/services/zoneminder/zm_manage_view_pane.py
@@ -1,6 +1,5 @@
 from typing import Dict
 
-from hi.apps.entity.models import Entity
 from hi.integrations.integration_data import IntegrationData
 from hi.integrations.integration_manage_view_pane import IntegrationManageViewPane
 
@@ -11,9 +10,4 @@ class ZmManageViewPane( IntegrationManageViewPane ):
         return 'zoneminder/panes/zm_manage.html'
 
     def get_template_context( self, integration_data : IntegrationData ) -> Dict[ str, object ]:
-
-        has_entities = Entity.objects.filter( integration_id = integration_data.integration_id ).exists()
-        
-        return {
-            'has_entities': has_entities,
-        }
+        return {}
